middlewareの修正

+++ routes/web.php
@@ -31,8 +31,8 @@ Route::get('/rooms', [App\Http\Controllers\RoomsController::class, 'index'])->na
 Route::get('/rooms/create', [App\Http\Controllers\RoomsController::class, 'create'])->middleware(['auth', 'verified', 'can:isAdmin'])->name('rooms.create');
 Route::post('/rooms', [App\Http\Controllers\RoomsController::class, 'store'])->middleware(['auth', 'verified', 'can:isAdmin'])->name('rooms.store');
 Route::get('/rooms/{id}', [App\Http\Controllers\RoomsController::class, 'show'])->name('rooms.show')->where('id', '[0-9]+');
-Route::get('/rooms/{id}/edit', [App\Http\Controllers\RoomsController::class, 'edit'])->name('rooms.edit')->where('id', '[0-9]+');
-Route::patch('/rooms/{id}/update', [App\Http\Controllers\RoomsController::class, 'update'])->name('rooms.update')->where('id', '[0-9]+');
-Route::delete('/rooms/{id}', [App\Http\Controllers\RoomsController::class, 'delete'])->name('rooms.destroy')->where('id', '[0-9]+');
+Route::get('/rooms/{id}/edit', [App\Http\Controllers\RoomsController::class, 'edit'])->middleware(['auth', 'verified', 'can:isAdmin'])->name('rooms.edit')->where('id', '[0-9]+');
+Route::patch('/rooms/{id}/update', [App\Http\Controllers\RoomsController::class, 'update'])->middleware(['auth', 'verified', 'can:isAdmin'])->name('rooms.update')->where('id', '[0-9]+');
+Route::delete('/rooms/{id}', [App\Http\Controllers\RoomsController::class, 'delete'])->middleware(['auth', 'verified', 'can:isAdmin'])->name('rooms.destroy')->where('id', '[0-9]+');
 
 require __DIR__.'/auth.php';
